// src/app/modules/configuration/class-schedules/model/class-periods-multi-view.ts
/**
 * Course
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * ClassPeriodsMultiView
 */
export interface ClassPeriodsMultiView { 
    /**
     * Gets or Sets Id
     */
    id?: number;
    /**
     * Gets or Sets CourseId
     */
    courseId?: number;
    /**
     * Gets or Sets CourseId
     */
    staffId?: number;
}