// src/app/modules/configuration/class-schedules/model/class-timetable-update-view.ts
/**
 * Course
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * ClassTimetableUpdateView
 */
export interface ClassTimetableUpdateView {
    /**
     * Gets or sets StartDate
     */
    startDate?: Date;
    /**
     * Gets or sets EndDate
     */
    endDate?: Date;
    id?: string;
}